・[add] タブグループ機能を追加
・[add] ダークモードに対応した
・[add] APIフックを使うかどうかはオプションで選べるようにした
・[add] "同名のタブが存在する時、タブの親フォルダ名を同時に表示する" オプションを追加
・[add] エクスプローラーの一列選択表示の有無に関わらず、タブ切り替えでスクロール位置の復元ができるようになった

・[change] CreateProcessはAPIフックしないようにした
・[change] タブリストの保存形式をxmlからjsonへ変更(今バージョンでは自動的にjsonに変換して保存するが、次回のバージョンからは自動変換機能は消す予定)
・[change] 最近閉じたタブの保存形式をxmlからjsonへ変更(以下同文)
・[change] タブリストの保存は一時ファイルを経由して行うようにした(運が悪いとタブリストが消滅する事への対策)
(アイコンと詳細表示に限る)
・[change] リストビューにおいて、移動キー入力でいつもサムネイルツールチップが表示されていたのを、サムネイルツールチップが表示されている時だけ選択画像を切り替えて表示するようにした
・[change] インストールやアンインストール時にエクスプローラー関連のレジストリキーを変更するのをやめた
・[change] オプションのデフォルト設定値をいくつか変更

・[fix] リストビューの一列表示有効時、戻るボタン押しながらホイールスクロールで画像切り替えを実行した場合、マウスホバーでサムネイルツールチップが表示されなくなるバグを修正
・[fix] タブの大きさを指定するのチェックが外れている時、タブの高さの計算がおかしかったのを修正
・[fix] サムネイルツールチップを無効にしているのに、ホイールクリックでサムネイルツールチップが表示されてしまうのを修正

・[misc] 開発環境をVisualStudio2019へ移行
・[misc] コンパイラのwarningに対処した
・[misc] windows7関連のコードを削除

+++ ExpTabBar/ThumbnailTooltip.cpp
@@ -145,7 +145,7 @@ void	CThumbnailTooltip::_StartCreateImageDataThread()
 				CCritSecLock	lock(m_cs);
 				// キャッシュのサイズを超えたので削除
 				if (m_bAddImageCached == false &&
-					m_ImageCache.size() > CThumbnailTooltipConfig::s_nMaxThumbnailCache)
+					static_cast<int>(m_ImageCache.size()) > CThumbnailTooltipConfig::s_nMaxThumbnailCache)
 				{
 					auto& seqList = m_ImageCache.get<seq>();
 					seqList.erase(seqList.begin());
